backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
alembic==1.13.3
pymysql==1.1.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
jinja2==3.1.4
pytest==8.3.3
httpx==0.27.2
openpyxl==3.1.5
